Refactor: NewsletterController NewsletterRecomendationService로 직접 의존하도록 변경
